requirements.txt
krpc
pytest
pyyaml
